requirements.txt
Django==1.11.1
simplejson==3.12.0
boto3==1.4.7
djangorestframework==3.7.3
djangorestframework-xml==1.3.0
geopy==1.11.0